+++ R/helpers_ggbetween_t_subtitles.R
@@ -115,7 +115,7 @@ subtitle_t_parametric <- function(data,
 
   # effect size object
   effsize_df <-
-    effect_t_parametric(
+    effsize_t_parametric(
       formula = y ~ x,
       data = data,
       paired = paired,
@@ -585,26 +585,35 @@ subtitle_t_bayes <- function(data,
 }
 
 
-#' @title Calculating Cohen or Hedges (between-/within- or one
+#' @title Calculating Cohen's *d* or Hedge's *g* (for between-/within- or one
 #'   sample designs).
-#' @name effect_t_parametric
+#' @name effsize_t_parametric
 #' @author Chuck Powell
 #'
-#' @param formula this function only accepts the variables in
+#' @param formula This function only accepts the variables in
 #'   `formula` format e.g. `sleep_rem ~ vore` or `~ vore`.
-#' @param data a data frame or tibble containing the data
-#' @param mu if conducting a single sample test against a mean (Default: `0`).
-#' @param paired Logical indicating whether to treat the data as a oaired sample
-#'   e.g. `post ~ pre`.
+#' @param mu If conducting a single sample test against a mean (Default: `0`).
 #' @param hedges.correction Logical indicating whether to apply Hedges correction,
 #'   Hedge's *g* (Default: `TRUE`).
-#' @param conf.level A scalar value between 0 and 1. If unspecified, the
-#'    default is to return `95%` lower and upper confidence intervals (`0.95`).
 #' @param noncentral Logical indicating whether to use non-central
 #'   *t*-distributions for computing the confidence intervals (Default: `TRUE`).
+#' @inheritParams subtitle_t_parametric
 #'
 #' @importFrom stats t.test na.omit cor qt pt uniroot
 #' @importFrom tibble tibble
+#'
+#' @details
+#' This function is a rewrite of functionality provided in `lsr::cohensD` and
+#' `effsize::cohen.d`.
+#'
+#' References-
+#' \itemize{
+#'   \item Cooper, Harris, Hedges, Larry V., Valentine, Jeffrey C., *The Handbook of Research Synthesis and Meta-Analysis*, 2009.
+#'   \item Cumming, G., Finch, S., *A Primer On The Understanding, Use, And Calculation Of Confidence Intervals That Are Based On Central And Noncentral Distributions*, Educational and Psychological Measurement, Vol. 61 No. 4,
+#' August 2001 532-574.
+#'   \item Cohen, J. (1988). *Statistical power analysis for the behavioral sciences (2nd ed.)* Hillsdale, NJ: Lawrence Erlbaum Associates.
+#'   \item David C. Howell (2010). *Confidence Intervals on Effect Size* retrieved from (\url{https://www.uvm.edu/~dhowell/methods7/Supplements/Confidence\%20Intervals\%20on\%20Effect\%20Size.pdf}).
+#' }
 #'
 #' @examples
 #'
@@ -615,13 +624,13 @@ subtitle_t_bayes <- function(data,
 #' )
 #'
 #' # with defaults
-#' effect_t_parametric(
+#' effsize_t_parametric(
 #'   formula = sleep_rem ~ vore,
 #'   data = msleep_short,
 #' )
 #'
 #' # changing defaults
-#' effect_t_parametric(
+#' effsize_t_parametric(
 #'   formula = sleep_rem ~ vore,
 #'   data = msleep_short,
 #'   mu = 1, # ignored in this case
@@ -633,13 +642,13 @@ subtitle_t_bayes <- function(data,
 #' @export
 
 # function body
-effect_t_parametric <- function(formula = NULL,
-                                data = NULL,
-                                mu = 0,
-                                paired = FALSE,
-                                hedges.correction = TRUE,
-                                conf.level = .95,
-                                noncentral = TRUE) {
+effsize_t_parametric <- function(formula = NULL,
+                                 data = NULL,
+                                 mu = 0,
+                                 paired = FALSE,
+                                 hedges.correction = TRUE,
+                                 conf.level = 0.95,
+                                 noncentral = TRUE) {
   # -------------- input checking -------------------
 
   if (!is(formula, "formula") | !is(data, "data.frame")) {
@@ -666,17 +675,25 @@ effect_t_parametric <- function(formula = NULL,
     Z <- -qt((1 - conf.level) / 2, df)
     lower.ci <- c(d - Z * sqrt(sd(x)))
     upper.ci <- c(d + Z * sqrt(sd(x)))
-    tobject <- t.test(x, mu = mu, var.equal = TRUE, conf.level = conf.level)
+    tobject <-
+      stats::t.test(
+        x = x,
+        alternative = "two.sided",
+        mu = mu,
+        var.equal = TRUE,
+        na.action = na.omit,
+        conf.level = conf.level
+      )
     tvalue <- tobject$statistic
     dfvalue <- tobject$parameter
     civalue <- attr(tobject$conf.int, "conf.level")
     twosamples <- FALSE
   }
 
-
   # ---------------two independent samples by factor -------------------
 
-  if (length(formula) == 3 & !isTRUE(paired)) { # two samples by factor
+  # two samples by factor
+  if (length(formula) == 3 & !isTRUE(paired)) {
     outcome <- eval(formula[[2]], data)
     group <- eval(formula[[3]], data)
     group <- factor(group)
@@ -702,21 +719,32 @@ effect_t_parametric <- function(formula = NULL,
     lower.ci <- c(d - Z * Sigmad)
     upper.ci <- c(d + Z * Sigmad)
     method <- "Cohen's d"
-    tobject <- t.test(x, y, var.equal = TRUE, conf.level = conf.level)
+    tobject <-
+      stats::t.test(
+        x = x,
+        y = y,
+        var.equal = TRUE,
+        alternative = "two.sided",
+        na.action = na.omit,
+        conf.level = conf.level
+      )
     tvalue <- tobject$statistic
     dfvalue <- tobject$parameter
     civalue <- attr(tobject$conf.int, "conf.level")
     twosamples <- TRUE
   }
 
-
   # -------------- two paired samples in matching columns -------------------
 
   if (length(formula) == 3 & isTRUE(paired)) {
-    if (is.factor(eval(formula[[3]], data))) {
+    if (is.factor(eval(formula[[3]], data)) ||
+        is.character(eval(formula[[3]], data))) {
       outcome <- eval(formula[[2]], data)
       group <- eval(formula[[3]], data)
-      group <- droplevels(group)
+      group <- droplevels(as.factor(group))
+      if (nlevels(group) != 2L) {
+        stop("grouping factor must have exactly 2 levels")
+      }
       x <- split(outcome, group)
       y <- x[[2]]
       x <- x[[1]]
@@ -742,7 +770,15 @@ effect_t_parametric <- function(formula = NULL,
     upper.ci <- c(d + Z * Sigmad)
     method <- "Cohen's d"
     diffscores <- as.vector(y - x)
-    tobject <- t.test(diffscores, mu = 0, var.equal = TRUE, conf.level = conf.level)
+    tobject <-
+      stats::t.test(
+        x = diffscores,
+        alternative = "two.sided",
+        mu = 0,
+        var.equal = TRUE,
+        na.action = na.omit,
+        conf.level = conf.level
+      )
     tvalue <- tobject$statistic
     dfvalue <- tobject$parameter
     civalue <- attr(tobject$conf.int, "conf.level")
@@ -803,14 +839,24 @@ effect_t_parametric <- function(formula = NULL,
       method = method,
       estimate = d,
       conf.low = ncp.lower.ci,
-      conf.high = ncp.upper.ci
+      conf.high = ncp.upper.ci,
+      conf.level = conf.level,
+      alternative = "two.sided",
+      paired = paired,
+      noncentral = noncentral,
+      var.equal = TRUE
     ))
   } else {
     return(tibble(
       method = method,
       estimate = d,
       conf.low = lower.ci,
-      conf.high = upper.ci
+      conf.high = upper.ci,
+      conf.level = conf.level,
+      alternative = "two.sided",
+      paired = paired,
+      noncentral = noncentral,
+      var.equal = TRUE
     ))
   }
 }